requirements.txt
google-generativeai>=0.8.0
asyncio
dataclasses
typing